Download contrail-status image in openshift infra.
contrail-status binary downloads the contrail-status image when the
command is invoked (i.e not at the time of provision/install).
This implies that the docker credentials for download contrail-status
image should be configured on all nodes where contrail-status can/would
be run.

This commit addresses this requirement by downloading contrail-status
image at the time of provisioning so docker credentials are not required
to be configured in docker daemon.

Change-Id: I7a73ffa046b7c7e733668e877c3bf2150ef17d52
Closes-Bug: #1800299